+++ tests/OsType/ValueBuilders/BuildTypeFromLsbReleaseTest.php
@@ -291,6 +291,10 @@ class BuildTypeFromLsbReleaseTest extends PHPUnit_Framework_TestCase
                 new CentOS('7.2')
             ],
             [
+                __DIR__ . '/lsb_release-examples/debian-8.2.php',
+                new Debian('8.2')
+            ],
+            [
                 __DIR__ . '/lsb_release-examples/ubuntu-15.04.php',
                 new Ubuntu('15.04')
             ]